Changes for first critiques in Tri2/Sem2 2016
Quite a few changes in here, not individually tracked (sorry).

GitHub OAuth is re-added, needed for UQ's course. Unlike the Play 2.4 version, this does not rely on EssentialAction. Instead of importing handy-play-oauth, the code is added directly into the project.

Groups can now be allocated for critiques again.

"Is this by me" now takes account that it might be "by" a group that you are in.