Partial commit
Some task and crit tasks in T1 2017 involve combinations of Files, Videos, and Questionnaires.

Added Question types for these. Also added CompositeTasks. Not that support is not yet threaded through, and may need to change. For instance, currently it only makes sense to have CritiqueTask be the top-level TaskBody in a Task.