Wrap MarkupNode in a div to avoid replacement cascade
MarkupNode is very rarely "equal" because anonymous functions are not equal to each other. This was causing a problem, whereby the markup node's replacement also caused the video iframe below it to be replaced (because the diff algorithm discards elements until it finds a match for the right hand side)

Wrapping the markup node in an empty div solves this. The div is equal to itself, so the cascade does not occur. (Within the div, the markup node is still replaced, but that's in an element whose list only contains that child)